+++ lib/signers/rsa_signer.dart
@@ -78,7 +78,7 @@ class RSASigner implements Signer {
     _rsa.init(forSigning, params);
   }
 
-  RSASignature generateSignature(Uint8List message) {
+  RSASignature generateSignature(Uint8List message, {bool normalize = false}) {
     if (!_forSigning) {
       throw new StateError(
           "Signer was not initialised for signature generation");